Add files via upload
A small analysis program for images taken with the LASER spot and a reference grid. 
The image needs to have been transformed to FITS format (preferably the green color part)

The program read the fits-file, and fits an elliptic Gaussian to the LASER spot
It then subtracts the background (calculated from beyond an ellipse of "securityfac" times the 
fitted half lengths and widths and normalizes the content. 

The background-subtracted image is then integrated in concentric ellipses around the fitted center, 
and the image integral contained outside the ellipses calculated. 

The result is plot of half width/length vs. image integral, from which one can read off 
the beam diameters at the reference value of 1/e^2 (indicated by a horizontal line)